pyproject.toml
[project]
name = "cloudcosting"
version = "0.1.0"
description = "Multi-cloud infrastructure cost estimation tool"
authors = [{name = "Development Team"}]
readme = "README.md"
requires-python = ">=3.13"
license = "GPL-3.0-or-later"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pyyaml",
    "boto3",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "hypothesis",
    "ruff",
]

[project.scripts]
cloudcosting = "cloudcosting.cli:main"

[project.urls]
Homepage = "https://github.com/dawsonlp/cloudcosting"
Repository = "https://github.com/dawsonlp/cloudcosting"
Issues = "https://github.com/dawsonlp/cloudcosting/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cloudcosting"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "integration: tests that require real cloud API credentials",
]

[tool.ruff]
src = ["src"]
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["cloudcosting"]